Test1/todolist_test1.1/requirements.txt
Django == 4.2.0
djangorestframework
markdown
django-filter
django-cors-headers